+++ Modules/Intent.Modules.Application.DomainInteractions/Mapping/Resolvers/ServiceOperationMappingTypeResolver.cs
@@ -29,8 +29,6 @@ namespace Intent.Modules.Application.DomainInteractions.Mapping.Resolvers
             if (model.SpecializationType == "Stored Procedure Parameter" &&
                 model.TypeReference.Element?.SpecializationType == "Data Contract")
             {
-                _template.AddTypeSource("Domain.DataContract");
-                
                 return new ObjectInitializationMapping(mappingModel, _template);
             }