+++ Tests/EntityFrameworkCore.Repositories.TestApplication/EntityFrameworkCore.Repositories.TestApplication.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -6,6 +6,7 @@ using EntityFrameworkCore.Repositories.TestApplication.Application.Common.Interf
 using EntityFrameworkCore.Repositories.TestApplication.Domain.Common;
 using EntityFrameworkCore.Repositories.TestApplication.Domain.Common.Interfaces;
 using EntityFrameworkCore.Repositories.TestApplication.Domain.Contracts;
+using EntityFrameworkCore.Repositories.TestApplication.Domain.Contracts.MappableStoredProcs;
 using EntityFrameworkCore.Repositories.TestApplication.Domain.Entities;
 using EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Persistence.Configurations;
 using Intent.RoslynWeaver.Attributes;
@@ -26,6 +27,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Persis
         }
 
         public DbSet<SpResult> SpResults { get; set; }
+        public DbSet<EntityRecord> EntityRecords { get; set; }
 
         public DbSet<AggregateRoot1> AggregateRoot1s { get; set; }
         public DbSet<AggregateRoot2Composition> AggregateRoot2Compositions { get; set; }
@@ -59,6 +61,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Persis
 
             ConfigureModel(modelBuilder);
             modelBuilder.Entity<SpResult>().HasNoKey().ToView(null);
+            modelBuilder.Entity<EntityRecord>().HasNoKey().ToView(null);
             modelBuilder.ApplyConfiguration(new AggregateRoot1Configuration());
             modelBuilder.ApplyConfiguration(new AggregateRoot2CompositionConfiguration());
             modelBuilder.ApplyConfiguration(new AggregateRoot3AggCollectionConfiguration());